# FIX a bug in feature_extraction.text.strip_accents_unicode