@rth 2019-09-26
Thanks! Could you please benchmark `CountVectorizer` before and after this fix say on `datasets.fetch_20newsgroups()`?