@DGrady 2019-09-27
@rth sure. I think it would also be useful to look at benchmarks on data that is not primarily ASCII-compatible. I used the following methodology to create a set of "documents" by randomly sampling code points from certain ranges, and chose the number and length of the strings to be roughly the same as the newsgroups data. (I'm not trying to present this as a way to get realistic non-ASCII text; this is just to get a set of data that will not trigger the fast path in the string processing functions.)

```python
import random

ranges = (
    range(0x0020, 0x007e + 1), # Basic Latin
    range(0x00a0, 0x00ff + 1), # Latin supplement; accented Latin letters
    range(0x0400, 0x04ff + 1), # Cyrillic
    range(0x0600, 0x06ff + 1), # Arabic
    range(0x4e00, 0x62ff + 1), # Some CJK
)

code_points = list()
for r in ranges:
    code_points.extend(r)

def random_string(n):
    return "".join(chr(i) for i in random.choices(code_points, k=n))

random_corpus = [random_string(2_000) for _ in range(10_000)]

with open("random_corpus.txt", "w", encoding="UTF-8") as f:
    f.write("\n".join(random_corpus))
```

To benchmark, I used IPython's `%timeit` magic with the following methodology, running each case in a fresh interpreter:

```python
from pathlib import Path
from sklearn.datasets import fetch_20newsgroups
from sklearn.feature_extraction.text import CountVectorizer
newsgroups = fetch_20newsgroups()
random_corpus = Path("random_corpus.txt").read_text().splitlines()
%timeit -r50 CountVectorizer(strip_accents="unicode").fit(# CORPUS #)
```

where `# CORPUS #` is either `newsgroups.data` or `random_corpus`. The results were

| Case | Newsgroups | Random strings |
| --- | --- | --- |
| master | 1.87 s ± 183 ms per loop | 4.18 s ± 468 ms per loop |
| This PR | 3.59 s ± 219 ms per loop | 4.03 s ± 332 ms per loop |
| Variant 1 | 2.23 s ± 237 ms per loop | 4.63 s ± 463 ms per loop |
| Variant 2 | 1.85 s ± 273 ms per loop | 4.12 s ± 269 ms per loop |

Since the PR currently introduces a significant slowdown on the newsgroups data, I tried a couple of guards in the spirit of the existing code. Variant 1 is

```python
def strip_accents_unicode(s):
    if ord(max(s)) < 128:
        return s
    else:
        normalized = unicodedata.normalize('NFKD', s)
        return ''.join([c for c in normalized if not unicodedata.combining(c)])
```

and variant 2 is

```python
def strip_accents_unicode(s):
    try:
        s.encode("ASCII", errors="strict")
        return s
    except UnicodeEncodeError:
        normalized = unicodedata.normalize('NFKD', s)
        return ''.join([c for c in normalized if not unicodedata.combining(c)])
```

I have read that benchmarking things accurately is difficult, and I'm not sure how seriously to take these numbers. To the extent we believe them, they seem to indicate that none of these cases make much difference for random non-ASCII-compatible strings, but that it's useful to have a guard to fast-path ASCII-compatible strings.

I guess we should go with variant 2, if that looks all right to everyone?