@rth 2019-10-01
Thanks for the detailed benchmarks. In Python 3.7+ there is `str.isascii` that should be used instead of catching the exception I think.

Could you try,
```diff
diff --git a/sklearn/feature_extraction/text.py b/sklearn/feature_extraction/text.py
index bb5a9d646..b3862ce35 100644
--- a/sklearn/feature_extraction/text.py
+++ b/sklearn/feature_extraction/text.py
@@ -111,6 +111,16 @@ def _analyze(doc, analyzer=None, tokenizer=None, ngrams=None,
     return doc
 
 
+def _isascii(s):
+    # XXX: this function implements str.isascii from Python 3.7+
+    #      and can be removed once support for Python <=3.6 is dropped.
+    try:
+        s.encode("ASCII", errors="strict")
+        return True
+    except UnicodeEncodeError:
+        return False
+
+
 def strip_accents_unicode(s):
     """Transform accentuated unicode symbols into their simple counterpart
 
@@ -129,10 +139,17 @@ def strip_accents_unicode(s):
         Remove accentuated char for any unicode symbol that has a direct
         ASCII equivalent.
     """
-    normalized = unicodedata.normalize('NFKD', s)
-    if normalized == s:
+    if hasattr(s, "isascii"):
+        # Python 3.7+
+        ascii_only = s.isascii()
+    else:
+        ascii_only = _isascii(s)
+
+    if ascii_only:
         return s
     else:
+        normalized = unicodedata.normalize('NFKD', s)
+
         return ''.join([c for c in normalized if not unicodedata.combining(c)])
```
if it's slower, we can still go with your version 2 but with a comment saying that we should use `str.isascii` when support for Python 3.6 is dropped.