@rth 2019-10-03
> (I can't find a C API for unicodedata.combining right now).

Doesn't look like it's exposed in the public API?

https://github.com/python/cpython/search?q=unicodedata_UCD_combining_impl

and `rg "combining" /usr/include/python3.7m/` doesn't find anything.